Adjusting the ExperienceStore to make use of the assembler pattern used by the rest of the SDK.
Resolves #14
